requirements.txt
streamlit==1.37.1
streamlit-drawable-canvas==0.9.3
pillow==10.4.0
numpy==1.26.4
